requirements.txt
# 核心框架
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0

# API和工具
arxiv>=2.1.0
requests>=2.32.0
pypdf>=4.3.0

# Web界面
gradio>=4.44.0

# 数据处理
pydantic>=2.9.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# 异步支持
aiohttp>=3.10.0
asyncio>=3.4.3

# 日志和工具
tenacity>=9.0.0
tqdm>=4.66.0
